move rendering of summary texts
